pyutils/requirements.txt
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
idna==2.8
numpy==1.17.2
Pillow==6.1.0
requests==2.22.0
selenium==3.141.0
soupsieve==1.9.3
urllib3==1.25.3
